+++ exercises/03/after/Shipping/OrderPlacedHandler.cs
@@ -16,15 +16,4 @@ namespace Shipping
             return Task.CompletedTask;
         }
     }
-
-    public class OrderBilledHandler : IHandleMessages<OrderBilled>
-    {
-        static ILog log = LogManager.GetLogger<OrderBilledHandler>();
-
-        public Task Handle(OrderBilled message, IMessageHandlerContext context)
-        {
-            log.Info($"Received OrderBilled, OrderId = {message.OrderId} - Should we ship now?");
-            return Task.CompletedTask;
-        }
-    }
 }
